+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes between Rockpool releases will be documented in this file
 ### Changed
 ### Fixed
 * Remove dependence of imp module that was deprecated in Python 3.12
+* Fixed bug in multiplying `TSContinuous` objects with differing numbers of channels
 
 ### Deprecated
 ### Removed